Move logUnitDelta to StockBase.
